+++ renderlib/our_fft.h
@@ -1,5 +1,5 @@
 #include <complex>
 #include <vector>
 
-std::vector<std::complex<double> > ourFFT(const std::vector<double>& list);
-std::vector<double> ourIFFT(const std::vector<std::complex<double> >& rhs);
+std::vector<std::complex<float> > ourFFT(const std::vector<float>& list);
+std::vector<float> ourIFFT(const std::vector<std::complex<float> >& rhs);
